# Luminance calculation optimisation