--- a/autocrop.go
+++ b/autocrop.go
@@ -53,13 +53,24 @@ func BoundsForThreshold(img *image.NRGBA, energyThreshold float32) image.Rectang
 // Energies returns the total row and column energies for the specified region
 // of an image.
 func Energies(img *image.NRGBA, r image.Rectangle) (cols, rows []float32) {
+
+	// Need 1 pixel more luminance data on each side so that all energy
+	// calculations are for pixels with a full set of neighbours.
+	luminanceBounds := r
+	luminanceBounds.Min.X--
+	luminanceBounds.Min.Y--
+	luminanceBounds.Max.X++
+	luminanceBounds.Max.Y++
+
+	luminances, alphas := luminancesAndAlphas(img, luminanceBounds)
+
 	cols = make([]float32, r.Dx(), r.Dx())
 	rows = make([]float32, r.Dy(), r.Dy())
 
 	// Calculate total column and row energies
 	for i, row := r.Min.Y, 0; i < r.Max.Y; i, row = i+1, row+1 {
 		for j, col := r.Min.X, 0; j < r.Max.X; j, col = j+1, col+1 {
-			e := energy(img, j, i)
+			e := energy(luminances, alphas, luminanceBounds.Dx(), col+1, row+1)
 			cols[col] += e
 			rows[row] += e
 		}
@@ -85,22 +96,16 @@ func colourAt(img *image.NRGBA, x, y int) (col srgb.Color, alpha float32) {
 	return srgb.ColorFromNRGBA(img.NRGBAAt(x, y))
 }
 
-func energy(img *image.NRGBA, x, y int) float32 {
-	neighbours := [8]float32{
-		luminance(colourAt(img, x-1, y-1)),
-		luminance(colourAt(img, x, y-1)),
-		luminance(colourAt(img, x+1, y-1)),
-		luminance(colourAt(img, x-1, y)),
-		luminance(colourAt(img, x+1, y)),
-		luminance(colourAt(img, x-1, y+1)),
-		luminance(colourAt(img, x, y+1)),
-		luminance(colourAt(img, x+1, y+1)),
-	}
+func energy(luminances, alphas []float32, width int, x, y int) float32 {
+	center := y*width + x
 
-	eX := neighbours[0] + neighbours[3] + neighbours[5] - neighbours[2] - neighbours[4] - neighbours[7]
-	eY := neighbours[0] + neighbours[1] + neighbours[2] - neighbours[5] - neighbours[6] - neighbours[7]
+	// North west + west + south west - north east - east - south east
+	eX := luminances[center-width-1] + luminances[center-1] + luminances[center+width-1] - luminances[center-width+1] - luminances[center+1] - luminances[center+width+1]
 
-	return float32((math.Abs(float64(eX)) + math.Abs(float64(eY))) * (float64(img.NRGBAAt(x, y).A) / 255))
+	// North west + north + north east - south west - south - south east
+	eY := luminances[center-width-1] + luminances[center-width] + luminances[center-width+1] - luminances[center+width-1] - luminances[center+width] - luminances[center+width+1]
+
+	return float32((math.Abs(float64(eX)) + math.Abs(float64(eY))) * float64(alphas[center]))
 }
 
 func findFirstEnergyBound(energies []float32, maxEnergy, threshold float32) (bound int) {
@@ -146,4 +151,24 @@ func findMaxEnergy(energies []float32) float32 {
 
 func luminance(c srgb.Color, alpha float32) float32 {
 	return c.Luminance() + alpha
+}
+
+func luminancesAndAlphas(img *image.NRGBA, r image.Rectangle) (luminances, alphas []float32) {
+
+	luminances = make([]float32, r.Dx()*r.Dy(), r.Dx()*r.Dy())
+	alphas = make([]float32, r.Dx()*r.Dy(), r.Dx()*r.Dy())
+
+	index := 0
+
+	// Get the luminances and alphas for all pixels
+	for i := r.Min.Y; i < r.Max.Y; i++ {
+		for j := r.Min.X; j < r.Max.X; j++ {
+			c, a := colourAt(img, j, i)
+			luminances[index] = luminance(c, a)
+			alphas[index] = a
+			index++
+		}
+	}
+
+	return luminances, alphas
 }